Move `InteractiveSourcesAdder` to adders package
