+++ src/statue/cli/config/__init__.py
@@ -1,7 +1,7 @@
 """Configuration related CLIs."""
 from statue.cli.config.config_cli import config_cli
-from statue.cli.config.config_general import set_mode_cli, set_history_size_cli
 from statue.cli.config.config_commands import fixate_commands_versions_cli
+from statue.cli.config.config_general import set_history_size_cli, set_mode_cli
 from statue.cli.config.config_init import init_config_cli
 from statue.cli.config.config_show import show_config_cli, show_config_tree_cli
