+++ src/statue/cli/config/interactive_adders/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration components interactive adders."""
